Updated multiple components and fixed various issues fix: resolve calculation bugs, improve UX validation, and harden data integrity • Fix unit conversion roundtrip drift: replace roundToHalf with roundToQuarter to prevent permanent precision loss on repeated lb↔kg toggles (2.5→1.0→2.0) • Fix RecoveryAnalyzer rep drop comparison using wrong baseline • Fix WarmupEngine generating barbell-only warmups for dumbbell exercises; add dedicated dumbbellCompoundWarmups path • Fix displayWeightStep in kg mode rounding to nearest 0.5 kg • Add achievedWeight to PlateLoading with residue warning when plates can’t exactly match the target weight • Disable template save when exercise list is empty (previously only validated non-empty name) • Add confirmationDialog to “Reset Inventory to Default” to prevent accidental data loss • Migrate weight storage from Int to Double across all models and stores • Add AnalyticsCache with Combine to avoid recomputing stats on every render • Add ExerciseRegistry to resolve exercises from both SeedData and templates • Debounce persistence writes with DispatchWorkItem • Add @MainActor to all ObservableObject stores • Fix deprecated onChange(of:) API to two-parameter closure form • Use stable content-based IDs for chart views instead of regenerating UUIDs

+++ MyWorkout/Data/SeedData.swift
@@ -325,8 +325,8 @@ struct SeedData {
 
         minReps: Int = 8,
         maxReps: Int = 10,
-        increaseAmount: Int = 5,
-        deloadAmount: Int = 5,
+        increaseAmount: Double = 5,
+        deloadAmount: Double = 5,
         stallLimit: Int = 3,
         exerciseType: ExerciseType = .compound,
         progressionStrategy: ProgressionStrategy = .doubleProgression
